COMP: Eliminate 64-bit compile warnings about conversion from size_t to unsigned int. None of the entities involved should ever approach the 4G 32-bit limit of an unsigned int, so static_cast should be safe for the foreseeable future.

+++ Cable/Parsers/cableFunctionType.cxx
@@ -133,7 +133,7 @@ void FunctionType::AddArgument(Type* argument, bool hasDefault, const char *name
 //----------------------------------------------------------------------------
 unsigned int FunctionType::GetNumberOfArguments() const
 {
-  return m_ArgumentTypeVector.size();
+  return static_cast<unsigned int>(m_ArgumentTypeVector.size());
 }
 
 //----------------------------------------------------------------------------
